ig: fix %check.
%check is not run with "set -e", so we need to ensure everything is correct
throughout the test.

Fixes: f102317bd06e ("ig: add package v0.25.0 (#7777)")

+++ SPECS/ig/ig.spec
@@ -1,7 +1,7 @@
 Summary:        The eBPF tool and systems inspection framework for Kubernetes, containers and Linux hosts.
 Name:           ig
 Version:        0.25.0
-Release:        1%{?dist}
+Release:        2%{?dist}
 License:        Apache 2.0 and GPL 2.0 for eBPF code
 Vendor:         Microsoft Corporation
 Distribution:   Azure Linux
@@ -34,16 +34,29 @@ mkdir -p "%{buildroot}/%{_bindir}"
 install -D -m0755 bin/build/ig %{buildroot}/%{_bindir}
 
 %check
-make gadgets-unit-tests
-ig_file=$(mktemp /tmp/ig-XXXXXX.out)
-sudo ./bin/build/ig trace exec --host > $ig_file &
-ig_pid=$!
-sleep inf &
-sleep_pid=$!
-kill $ig_pid
-kill $sleep_pid
-grep -P "${sleep_pid}\s+\d+\s+sleep" $ig_file
-rm $ig_file
+set -e
+set -o pipefail
+
+# Inspektor Gadget provides unit tests but they rely on several components which
+# are not present in the chroot used to build and test the package, among
+# others:
+# * runc: https://github.com/inspektor-gadget/inspektor-gadget/blob/3c8d1455525b/pkg/container-hook/tracer.go#L302
+# * dockerd: https://github.com/inspektor-gadget/inspektor-gadget/blob/3c8d1455525b/pkg/container-utils/testutils/docker.go#L67
+# Even if we recreate a proper testing environment, we will still have problems
+# as, for example, the path tested will be inside the chroot while ig reports
+# the full path from host point of view.
+# For all these reasons, we will skip the unit tests and rather run a small
+# integration test.
+# Moreover, Inspektor Gadget CI covers Azure Linux extensively:
+# https://github.com/inspektor-gadget/inspektor-gadget/pull/1186/commits/066bf618d158
+if [ -d /sys/kernel/debug/tracing ]; then
+	sleep inf &
+	sleep_pid=$!
+	./bin/build/ig snapshot process --host | grep -qP "sleep\s+${sleep_pid}"
+	kill $sleep_pid
+else
+	echo "Skipping ig check as prerequisites are not satisfied in the chroot"
+fi
 
 %files
 %license LICENSE
@@ -51,6 +64,9 @@ rm $ig_file
 %{_bindir}/ig
 
 %changelog
+* Tue Mar 14 2023 Francis Laniel <flaniel@linux.microsoft.com> - 0.25.0-2
+- Fix %check.
+
 * Tue Mar 14 2023 Francis Laniel <flaniel@linux.microsoft.com> - 0.25.0-1
-- Original version for CBL-Mariner
+- Original version for Azure Linux
 - License Verified
